+++ src/main/java/com/coolerpromc/easybrewing/screen/widget/ChangeCapabilityButton.java
@@ -3,6 +3,7 @@ package com.coolerpromc.easybrewing.screen.widget;
 import com.coolerpromc.easybrewing.block.entity.ItemBrewingStationBE;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 import java.util.Optional;
 import java.util.function.Supplier;
 import net.minecraft.client.MinecraftClient;
@@ -14,10 +15,10 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Direction;
 
 public class ChangeCapabilityButton extends ButtonWidget {
-    private final Direction direction;
+    private final ItemBrewingStationBE.RelativeSide direction;
     private ItemBrewingStationBE.Slot slot;
 
-    public ChangeCapabilityButton(int x, int y, int width, int height, Text message, PressAction onPress, Direction direction, ItemBrewingStationBE.Slot slot) {
+    public ChangeCapabilityButton(int x, int y, int width, int height, Text message, PressAction onPress, ItemBrewingStationBE.RelativeSide direction, ItemBrewingStationBE.Slot slot) {
         super(x, y, width, height, message, onPress, Supplier::get);
         this.direction = direction;
         this.slot = slot;
@@ -34,7 +35,7 @@ public class ChangeCapabilityButton extends ButtonWidget {
 
         if (this.hovered){
             List<Text> tooltips = new ArrayList<>();
-            tooltips.add(Text.translatable("screen.easybrewing.direction_" + direction.getName()));
+            tooltips.add(Text.translatable("screen.easybrewing.direction_" + direction.name().toLowerCase(Locale.ROOT)));
             tooltips.add(Text.translatable("screen.easybrewing.change_slot").formatted(Formatting.GRAY));
             guiGraphics.drawTooltip(MinecraftClient.getInstance().textRenderer, tooltips, Optional.empty(), mouseX, mouseY);
         }
@@ -48,7 +49,7 @@ public class ChangeCapabilityButton extends ButtonWidget {
         return slot;
     }
 
-    public Direction getDirection() {
+    public ItemBrewingStationBE.RelativeSide getDirection() {
         return direction;
     }
 }